Minor refactoring; add sound.
